Use `rackup` to start app
This is the more traditional way to use raw Rack.

Note that Rack is happy with any object that responds to the `call`
method. The `Proc` was a shorthand for making such an object, but a Ruby
class is more typical.

The `ru` in `config.ru` stands for "rack up". This is the config file
for a command called `rackup`, which is the most typical way to start up
a Rack application.

There are lots of blog posts about how to build Ruby web apps with only
Rack. One that seems decent:
http://1devblog.org/article/framework-less-ruby-app-rack-based